EB-642: Implement query suggestion: 06. Support for split data

+++ src/divbase_tools/tasks.py
@@ -1,12 +1,16 @@
 import dataclasses
 import logging
 import os
+import re
 from pathlib import Path
 
+import yaml
 from celery import Celery
 
+from divbase_tools.exceptions import NoVCFFilesFoundError
 from divbase_tools.queries import BCFToolsInput, BcftoolsQueryManager, run_sidecar_metadata_query
 from divbase_tools.s3_client import S3FileManager, create_s3_file_manager
+from divbase_tools.vcf_dimension_indexing import VCFDimensionIndexManager
 
 logger = logging.getLogger(__name__)
 
@@ -99,17 +103,31 @@ def bcftools_pipe_task(
         filter_string=tsv_filter,
     )
 
+    if "view -r" in command:
+        files_to_download = check_for_unnecessary_files_for_region_query(
+            bucket_name=bucket_name,
+            s3_file_manager=s3_file_manager,
+            command=command,
+            files_to_download=metadata_result.unique_filenames,
+        )
+
+        sample_and_filename_subset = [
+            entry for entry in metadata_result.sample_and_filename_subset if entry["Filename"] in files_to_download
+        ]
+    else:
+        sample_and_filename_subset = metadata_result.sample_and_filename_subset
+
     _ = download_vcf_files(
-        files_to_download=metadata_result.unique_filenames,
+        files_to_download=files_to_download,
         bucket_name=bucket_name,
         s3_file_manager=s3_file_manager,
     )
 
     bcftools_inputs = dataclasses.asdict(
         BCFToolsInput(
-            sample_and_filename_subset=metadata_result.sample_and_filename_subset,
+            sample_and_filename_subset=sample_and_filename_subset,
             sampleIDs=metadata_result.unique_sample_ids,
-            filenames=metadata_result.unique_filenames,
+            filenames=files_to_download,
         )
     )
 
@@ -120,10 +138,63 @@ def bcftools_pipe_task(
         return {"status": "error", "error": str(e), "task_id": task_id}
 
     upload_results_file(output_file=Path(output_file), bucket_name=bucket_name, s3_file_manager=s3_file_manager)
-    delete_job_files_from_worker(
-        vcf_paths=metadata_result.unique_filenames, metadata_path=metadata_path, output_file=output_file
-    )
+    delete_job_files_from_worker(vcf_paths=files_to_download, metadata_path=metadata_path, output_file=output_file)
     return {"status": "completed", "output_file": output_file, "submitter": user_name}
+
+
+@app.task(name="tasks.update_vcf_dimensions_task")
+def update_vcf_dimensions_task(bucket_name: str, user_name: str = "Default User"):
+    """
+    Update the VCF dimensions file for the specified bucket. Ensures that the index only covers files in the bucket at the time of task execution.
+
+    Re: vcfs_deleted_from_bucket_since_last_indexing: since both lists it is dependent on are already calculated,
+    set difference is a faster operation than a new list comp.
+    """
+    task_id = update_vcf_dimensions_task.request.id
+    s3_file_manager = create_s3_file_manager(url="http://minio:9000")
+
+    all_files = s3_file_manager.list_files(bucket_name=bucket_name)
+    vcf_files = [file for file in all_files if file.endswith(".vcf") or file.endswith(".vcf.gz")]
+
+    if not vcf_files:
+        raise NoVCFFilesFoundError(
+            f"VCF dimensions file could not be generated since no VCF files were found in the project: {bucket_name}. Please upload at least one VCF file and run this command again."
+        )
+
+    manager = VCFDimensionIndexManager(bucket_name=bucket_name, s3_file_manager=s3_file_manager)
+    already_indexed_vcfs = manager.get_indexed_filenames()
+
+    non_indexed_vcfs = [file for file in vcf_files if file not in already_indexed_vcfs]
+
+    _ = download_vcf_files(
+        files_to_download=non_indexed_vcfs,
+        bucket_name=bucket_name,
+        s3_file_manager=s3_file_manager,
+    )
+
+    files_indexed_by_this_job = []
+    for file in non_indexed_vcfs:
+        try:
+            manager.add_dimension_entry(vcf_filename=file)
+            files_indexed_by_this_job.append(file)
+        except Exception as e:
+            logger.error(f"Error in dimensions indexing task: {str(e)}")
+            return {"status": "error", "error": str(e), "task_id": task_id}
+
+    vcfs_deleted_from_bucket_since_last_indexing = list(set(already_indexed_vcfs) - set(vcf_files))
+
+    if vcfs_deleted_from_bucket_since_last_indexing:
+        for file in vcfs_deleted_from_bucket_since_last_indexing:
+            manager.remove_dimension_entry(vcf_filename=file)
+
+    delete_job_files_from_worker(vcf_paths=non_indexed_vcfs)
+    manager._upload_bucket_dimensions_file(dimensions_data=manager.dimensions_info)
+
+    return {
+        "status": "completed",
+        "submitter": user_name,
+        "VCF files that were added to dimensions file by this job": files_indexed_by_this_job,
+    }
 
 
 def download_sample_metadata(metadata_tsv_name: str, bucket_name: str, s3_file_manager: S3FileManager) -> Path:
@@ -159,7 +230,70 @@ def upload_results_file(output_file: Path, bucket_name: str, s3_file_manager: S3
     )
 
 
-def delete_job_files_from_worker(vcf_paths: list[Path], metadata_path: Path, output_file: Path) -> None:
+def check_for_unnecessary_files_for_region_query(
+    bucket_name: str, s3_file_manager: S3FileManager, command: str, files_to_download: list[str]
+) -> dict:
+    """
+    If the 'view -r' query is present, read the .vcf_dimensions.yaml file and check if the specified scaffolds are available in the VCF files.
+    If a file in files_to_download (as identified by sidecar sample metadata query) does not contain any of the specified scaffolds, it will be skipped.
+    This will save time and resources by avoiding unnecessary transfer of irrelevant files between the bucket and the worker.
+
+    NOTE! There is a risk that users mistype...
+
+    TODO ensure that the errors raised here are raised before submitting the task to celery.
+    """
+
+    try:
+        dimensions_index = read_vcf_dimensions_file(bucket_name=bucket_name, s3_file_manager=s3_file_manager)
+    except FileNotFoundError:
+        return files_to_download
+        # TODO for now, this makes it so that if the dimensions file is not present, the job continues by skipping the check of unnecessary files. There should at least be a message that tells this to the user
+
+    scaffolds = []
+    files_to_download_updated = []
+    matches = re.findall(r"view\s+-r\s+([^\s;]+)", command)
+    for match in matches:
+        scaffolds.extend([region.strip() for region in match.split(",") if region.strip()])
+    # TODO if there are no matches, return files to download
+    for file in files_to_download:
+        record = None
+        for rec in dimensions_index.get("dimensions", []):
+            if rec.get("filename") == file:
+                record = rec
+                break
+        if not record:
+            print(f"File '{file}' is not indexed in the VCF dimensions file.")
+            continue
+
+        record_scaffolds = set(record.get("dimensions", {}).get("scaffolds", []))
+        for scaffold in scaffolds:
+            if scaffold in record_scaffolds:
+                print(f"Scaffold '{scaffold}' is present in file '{file}'.")
+                if file not in files_to_download_updated:
+                    files_to_download_updated.append(file)
+
+    if files_to_download_updated == []:
+        raise ValueError(
+            "Based on the 'view -r' query and the VCF scaffolds indexed in DivBase, there are no VCF files in the project that fulfills the query. Please try another -r query with scaffolds/chromosomes that are present in the VCF files."
+        )
+        # TODO this stops jobs from executing. Which might not be intuitive. Such subsests would be able to run in bftools but the result would be an merged VCF with no variants (header only). Should this be a warning instead that will be returned in the worker results?
+
+    return files_to_download_updated
+
+    # TODO when there is a command implemented that reads the .vcf_dimensions.yaml file and list all scaffolds in the project, add that command as a help to this error message.
+
+
+def read_vcf_dimensions_file(bucket_name: str, s3_file_manager: S3FileManager) -> dict:
+    try:
+        content = s3_file_manager.download_s3_file_to_str(key=".vcf_dimensions.yaml", bucket_name=bucket_name)
+    except Exception as exc:
+        raise FileNotFoundError(f".vcf_dimensions.yaml not found in bucket '{bucket_name}'.") from exc
+    # TODO add a hint to the error telling how dimension indexing can be done.
+    # TODO use the get_dimensions_info instance method of the VCFDimensionIndexManager class instead of reading file diretly? don't want to download the file though
+    return yaml.safe_load(content)
+
+
+def delete_job_files_from_worker(vcf_paths: list[Path], metadata_path: Path = None, output_file: Path = None) -> None:
     """
     After uploading results to bucket, delete job files from the worker.
     """
@@ -169,13 +303,15 @@ def delete_job_files_from_worker(vcf_paths: list[Path], metadata_path: Path, out
             logger.info(f"deleted {vcf_path}")
         except Exception as e:
             logger.warning(f"Could not delete input VCF file from worker {vcf_path}: {e}")
-    try:
-        os.remove(metadata_path)
-        logger.info(f"deleted {metadata_path}")
-    except Exception as e:
-        logger.warning(f"Could not delete metadata file from worker {metadata_path}: {e}")
-    try:
-        os.remove(output_file)
-        logger.info(f"deleted {output_file}")
-    except Exception as e:
-        logger.warning(f"Could not delete output file from worker {output_file}: {e}")
+    if metadata_path is not None:
+        try:
+            os.remove(metadata_path)
+            logger.info(f"deleted {metadata_path}")
+        except Exception as e:
+            logger.warning(f"Could not delete metadata file from worker {metadata_path}: {e}")
+    if output_file is not None:
+        try:
+            os.remove(output_file)
+            logger.info(f"deleted {output_file}")
+        except Exception as e:
+            logger.warning(f"Could not delete output file from worker {output_file}: {e}")
